* split test_parsyfiles into two (old/new) * added a can_chain attribute to converters to prevent useless chains explicitly (used for primitive converters). * fixed doc in parsing_core * fixed reading xls files using pandas : streaming mode was not working because encoding was managed by the framework (too early) > now delegating to pandas * added first section in readme about collection parsing
