+++ parsyfiles/test_data/demo/simple_collection_inference/b.txt
@@ -1,0 +1,2 @@
+a,b,c,d
+1,2,3,4
