+++ parsyfiles/test_data/demo/simple_collection_inference/v.txt
@@ -1,0 +1,4 @@
+x = 1
+y = 1
+op = +
+expected_result = 2
