+++ parsyfiles/test_data/demo/simple_collection_inference/w.properties
@@ -1,0 +1,4 @@
+x=0
+y=1
+op=+
+expected_result=1
