+++ parsyfiles/test_data/demo/simple_collection_inference/y.cfg
@@ -1,0 +1,5 @@
+[main]
+x=1
+y=1
+op=-
+expected_result=0
