+++ parsyfiles/test_data/demo/simple_objects/test_diff_1.cfg
@@ -1,0 +1,5 @@
+[main]
+x=1
+y=1
+op=-
+expected_result=0
