+++ parsyfiles/test_data/demo/simple_objects/test_diff_2.ini
@@ -1,0 +1,5 @@
+[main]
+x=0
+y=1
+op=-
+expected_result=-1
