+++ parsyfiles/test_data/demo/simple_objects/test_diff_3_csv_format.txt
@@ -1,0 +1,2 @@
+x,y,op,expected_result
+5,4,-,1
