+++ parsyfiles/test_data/demo/simple_objects/test_sum_1.json
@@ -1,0 +1,6 @@
+{
+	"x": 1,
+	"y": 2,
+	"op": "+",
+	"expected_result": 3
+}
