+++ parsyfiles/test_data/demo/simple_objects/test_sum_2.properties
@@ -1,0 +1,4 @@
+x=0
+y=1
+op=+
+expected_result=1
