+++ parsyfiles/test_data/demo/simple_objects/test_sum_3_properties_format.txt
@@ -1,0 +1,4 @@
+x = 1
+y = 1
+op = +
+expected_result = 2
