+++ parsyfiles/test_data/demo/simple_objects/test_sum_4.yaml
@@ -1,0 +1,5 @@
+# a YAML dictionary
+x: 2
+y: 5
+op: +
+expected_result: 7
